requirements.txt
asgiref==3.11.1
bcrypt==5.0.0
Django==6.0.2
djangorestframework==3.16.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
